requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.0.12
colorlog==6.6.0
croniter==1.3.7
frozenlist==1.3.0
hikari==2.0.0.dev110
hikari-lightbulb==2.2.4
hikari-miru==1.1.2
idna==3.3
mariadb==1.0.11
multidict==6.0.2
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
types-croniter==1.3.2
uvloop==0.16.0
yarl==1.7.2
